尝试把获取投影机15角点的过程变为手动框选，并重新修正屏幕比率phco.txt相关代码，目前没有彻底改好
